[qtconnectivity] MER: Change OBEX client interface to obexd 0.48. Fixes MER#1224.
OBEX client interface changed from org.openobex to org.bluez.obex
along with changes how sessions are acquired and transfers are
initiated in obexd 0.47. Code modified to reflect the changes.

As in Mer obexd 0.48 is used it's enough to have code that works with
one interface only; upstream code would need to work with both
variants. Thus this commit should not be upstreamed as such.